# backend/internal/api/routes.go
package api

import (
	"context"
	"log/slog"
	"net/http"

	"github.com/chrismott/miniclass/internal/api/handlers"
	"github.com/chrismott/miniclass/internal/api/problems"
	"github.com/chrismott/miniclass/internal/auth"
	"github.com/danielgtaylor/huma/v2"
	"github.com/danielgtaylor/huma/v2/adapters/humachi"
	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/go-chi/cors"
)

const apiBasePath = "/api"

// RouterOptions configures the HTTP router independently of process startup.
type RouterOptions struct {
	AllowedOrigins         []string
	Database               handlers.DatabasePinger
	Logger                 *slog.Logger
	TrustedProxyCIDRs      []string
	Version                string
	Identity               auth.AccountResolver
	Claimer                handlers.InvitationClaimer
	Administrators         handlers.AdministratorManager
	InvitationClaimBaseURL string
	SchoolYears            handlers.SchoolYearService
	AuditLog               handlers.AuditLogReader
	Verifier               auth.Verifier
}

// NewRouter builds the complete API router and middleware chain. Routes are
// declared once through Huma so the same definitions produce the runtime
// handlers and the OpenAPI contract.
func NewRouter(options RouterOptions) chi.Router {
	router, _ := newRouter(options)
	return router
}

// NewOpenAPI builds the API declaration and returns its generated OpenAPI
// document for the cmd/openapi generator.
func NewOpenAPI(options RouterOptions) *huma.OpenAPI {
	_, api := newRouter(options)
	return api.OpenAPI()
}

func newRouter(options RouterOptions) (chi.Router, huma.API) {
	router := chi.NewRouter()
	logger := options.Logger
	if logger == nil {
		logger = slog.Default()
	}

	router.Use(
		middleware.RequestID,
		TrustedProxyRealIP(options.TrustedProxyCIDRs...),
		RequestLogger(logger),
		Recoverer(logger),
		cors.Handler(cors.Options{
			AllowedOrigins:   allowedOrigins(options.AllowedOrigins),
			AllowedMethods:   []string{"GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS"},
			AllowedHeaders:   []string{"Accept", "Authorization", "Content-Type", "X-CSRF-Token"},
			AllowCredentials: false,
			MaxAge:           300,
		}),
	)

	router.NotFound(func(w http.ResponseWriter, _ *http.Request) {
		problems.Write(w, problems.New(http.StatusNotFound, problems.RouteNotFound, "route not found"))
	})
	router.MethodNotAllowed(func(w http.ResponseWriter, _ *http.Request) {
		problems.Write(w, problems.New(http.StatusMethodNotAllowed, problems.MethodNotAllowed, "method not allowed"))
	})

	version := options.Version
	if version == "" {
		version = defaultServerVersion
	}
	options.Version = version
	api := humachi.New(router, humaConfig(version))
	api.UseMiddleware(auth.Middleware(options.Verifier, options.Identity, writeAuthError))
	registerOperations(api, options)
	addProblemTypesToContract(api.OpenAPI())
	return router, api
}

func humaConfig(version string) huma.Config {
	config := huma.DefaultConfig("MiniClass API", version)
	// The committed contract is generated by cmd/openapi. Keep the runtime
	// surface limited to the application routes until an explicit docs route is
	// designed.
	config.OpenAPIPath = ""
	config.DocsPath = ""
	config.SchemasPath = ""
	return config
}

func registerOperations(api huma.API, options RouterOptions) {
	registerOperation(api, huma.Operation{
		OperationID: "get-api-root",
		Method:      http.MethodGet,
		Path:        apiBasePath,
		Summary:     "Get API information",
	}, auth.CapabilityAuthenticated, false, apiRoot)
	registerOperation(api, huma.Operation{
		OperationID: "get-api-root-trailing-slash",
		Method:      http.MethodGet,
		Path:        apiBasePath + "/",
		Summary:     "Get API information",
	}, auth.CapabilityAuthenticated, false, apiRoot)

	health := handlers.NewHealthHandler(options.Database, options.Version)
	registerOperation(api, huma.Operation{
		OperationID: "get-health",
		Method:      http.MethodGet,
		Path:        apiBasePath + "/health",
		Summary:     "Check API and database health",
		Errors:      []int{http.StatusServiceUnavailable},
	}, auth.CapabilityAuthenticated, false, health.Handle)

	registerOperation(api, huma.Operation{
		OperationID: "get-me",
		Method:      http.MethodGet,
		Path:        apiBasePath + "/me",
		Summary:     "Get the authenticated principal",
	}, auth.CapabilityAuthenticated, false, (handlers.MeHandler{}).Handle)

	auditLog := handlers.NewAuditLogHandler(options.AuditLog)
	registerOperation(api, huma.Operation{
		OperationID: "get-audit-log",
		Method:      http.MethodGet,
		Path:        apiBasePath + "/audit-log",
		Summary:     "Read the audit log",
		Errors:      []int{http.StatusBadRequest, http.StatusInternalServerError},
	}, auth.CapabilityReadAuditLog, false, auditLog.Handle)

	claim := handlers.NewClaimInvitationHandler(options.Claimer)
	registerOperation(api, huma.Operation{
		OperationID: "post-auth-claim-invitation",
		Method:      http.MethodPost,
		Path:        apiBasePath + "/auth/claim",
		Summary:     "Claim an administrator invitation",
		Errors:      []int{http.StatusBadRequest, http.StatusForbidden},
	}, auth.CapabilityAuthenticated, true, claim.Handle)

	administrators := handlers.NewAdministratorHandler(options.Administrators, options.InvitationClaimBaseURL)
	registerOperation(api, huma.Operation{
		OperationID: "get-administrators", Method: http.MethodGet, Path: apiBasePath + "/administrators",
		Summary: "List organization administrators",
	}, auth.CapabilityManageAdministrators, false, administrators.List)
	registerOperation(api, huma.Operation{
		OperationID: "post-administrator-invitation", Method: http.MethodPost, Path: apiBasePath + "/administrators",
		Summary: "Invite an administrator",
		Errors:  []int{http.StatusBadRequest, http.StatusConflict},
	}, auth.CapabilityManageAdministrators, false, administrators.Invite)
	registerOperation(api, huma.Operation{
		OperationID: "patch-administrator-role", Method: http.MethodPatch, Path: apiBasePath + "/administrators/{memberID}",
		Summary: "Change an administrator role",
		Errors:  []int{http.StatusBadRequest, http.StatusConflict, http.StatusNotFound},
	}, auth.CapabilityManageAdministrators, false, administrators.ChangeRole)
	registerOperation(api, huma.Operation{
		OperationID: "delete-administrator", Method: http.MethodDelete, Path: apiBasePath + "/administrators/{memberID}",
		Summary: "Remove an administrator",
		Errors:  []int{http.StatusConflict, http.StatusNotFound},
	}, auth.CapabilityManageAdministrators, false, administrators.Remove)
	registerOperation(api, huma.Operation{
		OperationID: "post-administrator-invitation-resend", Method: http.MethodPost, Path: apiBasePath + "/administrators/{memberID}/invitation/resend",
		Summary: "Regenerate an administrator invitation",
		Errors:  []int{http.StatusConflict, http.StatusNotFound},
	}, auth.CapabilityManageAdministrators, false, administrators.Resend)
	registerOperation(api, huma.Operation{
		OperationID: "post-administrator-invitation-revoke", Method: http.MethodPost, Path: apiBasePath + "/administrators/{memberID}/invitation/revoke",
		Summary: "Revoke an administrator invitation",
		Errors:  []int{http.StatusConflict, http.StatusNotFound},
	}, auth.CapabilityManageAdministrators, false, administrators.Revoke)

	schoolYears := handlers.NewSchoolYearHandler(options.SchoolYears)
	registerOperation(api, huma.Operation{
		OperationID: "list-school-years",
		Method:      http.MethodGet,
		Path:        apiBasePath + "/school-years",
		Summary:     "List school years",
	}, auth.CapabilityManageSchoolYear, false, schoolYears.List)
	registerOperation(api, huma.Operation{
		OperationID: "create-school-year",
		Method:      http.MethodPost,
		Path:        apiBasePath + "/school-years",
		Summary:     "Create a school year",
		Errors:      []int{http.StatusBadRequest, http.StatusConflict},
	}, auth.CapabilityManageSchoolYear, false, schoolYears.Create)
	registerOperation(api, huma.Operation{
		OperationID: "get-school-year",
		Method:      http.MethodGet,
		Path:        apiBasePath + "/school-years/{schoolYearID}",
		Summary:     "Get a school year",
		Errors:      []int{http.StatusNotFound},
	}, auth.CapabilityManageSchoolYear, false, schoolYears.Get)
	registerOperation(api, huma.Operation{
		OperationID: "update-school-year",
		Method:      http.MethodPatch,
		Path:        apiBasePath + "/school-years/{schoolYearID}",
		Summary:     "Edit or transition a school year",
		Errors:      []int{http.StatusBadRequest, http.StatusNotFound, http.StatusConflict},
	}, auth.CapabilityManageSchoolYear, false, schoolYears.Update)
	registerOperation(api, huma.Operation{
		OperationID: "delete-school-year",
		Method:      http.MethodDelete,
		Path:        apiBasePath + "/school-years/{schoolYearID}",
		Summary:     "Delete a school year",
		Errors:      []int{http.StatusNotFound, http.StatusConflict},
	}, auth.CapabilityManageSchoolYear, false, schoolYears.Delete)
}

func registerOperation[I, O any](api huma.API, operation huma.Operation, capability auth.Capability, allowUnresolved bool, handler func(context.Context, *I) (*O, error)) {
	if operation.Extensions == nil {
		operation.Extensions = map[string]any{}
	}
	if operation.Metadata == nil {
		operation.Metadata = map[string]any{}
	}
	operation.Extensions[auth.RequiredCapabilityExtension] = string(capability)
	operation.Metadata[auth.RequiredCapabilityExtension] = string(capability)
	if allowUnresolved {
		operation.Extensions[auth.AllowUnresolvedPrincipalExtension] = true
		operation.Metadata[auth.AllowUnresolvedPrincipalExtension] = true
	}
	operation.Security = []map[string][]string{{"bearerAuth": {}}}
	huma.Register(api, operation, handler)
}

type apiRootOutput struct {
	Body struct {
		Service string `json:"service" doc:"Name of the API service."`
	}
}

func apiRoot(context.Context, *struct{}) (*apiRootOutput, error) {
	response := &apiRootOutput{}
	response.Body.Service = "miniclass-api"
	return response, nil
}

func addProblemTypesToContract(openAPI *huma.OpenAPI) {
	if openAPI.Extensions == nil {
		openAPI.Extensions = map[string]any{}
	}
	definitions := problems.Definitions()
	problemSlugs := make([]string, 0, len(definitions))
	for _, definition := range definitions {
		problemSlugs = append(problemSlugs, string(definition.Slug))
	}
	openAPI.Extensions["x-miniclass-problem-types"] = problemSlugs
	if openAPI.Components != nil {
		if openAPI.Components.SecuritySchemes == nil {
			openAPI.Components.SecuritySchemes = map[string]*huma.SecurityScheme{}
		}
		openAPI.Components.SecuritySchemes["bearerAuth"] = &huma.SecurityScheme{
			Type: "http", Scheme: "bearer", BearerFormat: "JWT", Description: "Verified Supabase or local ES256/RS256 bearer token.",
		}
	}

	if openAPI.Components == nil || openAPI.Components.Schemas == nil {
		return
	}
	schemas := openAPI.Components.Schemas.Map()
	schemas["ProblemType"] = problems.Schema()
	if errorSchema, ok := schemas["ErrorModel"]; ok && errorSchema.Properties != nil {
		errorSchema.Properties["type"] = &huma.Schema{Ref: "#/components/schemas/ProblemType"}
	}
}

func writeAuthError(ctx huma.Context, failure auth.Failure) {
	slug := problems.AuthenticationUnavailable
	switch failure.Code {
	case auth.FailureAuthenticationRequired:
		slug = problems.AuthenticationRequired
	case auth.FailureInvalidToken:
		slug = problems.InvalidToken
	case auth.FailureNoOrganization:
		slug = problems.NoOrganization
	case auth.FailureMultipleOrganizations:
		slug = problems.MultipleOrganizations
	case auth.FailureCapabilityRequired:
		slug = problems.CapabilityRequired
	case auth.FailureMissingCapability:
		slug = problems.CapabilityNotDeclared
	}
	problems.WriteContext(ctx, problems.New(failure.Status, slug, failure.Detail))
}

func allowedOrigins(origins []string) []string {
	if len(origins) == 0 {
		return []string{"*"}
	}
	return origins
}